work on reorganizing outputs
